requirements.txt
anyio==4.4.0
APScheduler==3.6.3
cachetools==4.2.2
certifi==2024.2.2
charset-normalizer==3.3.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
imgurpython==1.1.7
pillow==10.3.0
pypng==0.20220715.0
python-telegram-bot==13.13
pytz==2024.1
qrcode==7.4.2
requests==2.32.3
six==1.16.0
sniffio==1.3.1
tornado==6.4
typing_extensions==4.12.0
tzlocal==5.2
urllib3==2.2.1
